Move all the menu staff to lib
